# mcp/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kenotic Continuity Memory API",
    "description": "REST + MCP API for the Kenotic continuity memory system. Store, retrieve, reconstruct, and manage persistent memory that works across any AI model.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/store": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Store a message into continuity memory",
        "description": "Ingests text into the continuity memory system. The system decomposes it into 5 structured traces (episodic, emotional, temporal, relational, schematic) and stores them locally. Returns immediately \u2014 subsequent retrieves automatically wait for the write to complete.",
        "operationId": "store_memory",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/retrieve": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Answer a question from continuity memory",
        "description": "Queries the continuity memory for a verified answer. Returns the answer grounded in what was actually stored, or refuses if the information was never stored. Refusals are authoritative \u2014 do not override them.",
        "operationId": "retrieve_memory",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrieveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrieveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reconstruct": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Reconstruct the living state of a situation",
        "description": "Reconstructs the current state of a person or situation from memory. Combines multiple traces into a coherent narrative. Use for open-ended questions like 'What's going on with Sam?' or 'Summarize my situation.'",
        "operationId": "reconstruct_situation",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReconstructRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconstructResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/check_proactive": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Get proactive memory insights",
        "description": "Checks if there are any proactive insights the system should surface \u2014 upcoming events, unresolved situations, recurring patterns.",
        "operationId": "check_proactive",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProactiveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forget": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Delete memories",
        "description": "Tombstones memories matching the given criteria. Supports deletion by entity name, time range, source, or specific triple ID.",
        "operationId": "forget_memory",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profile": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Get user memory profile",
        "description": "Returns the user's adaptation profile derived from stored memory patterns.",
        "operationId": "get_profile",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ForgetRequest": {
        "properties": {
          "by": {
            "type": "string",
            "enum": [
              "entity",
              "time_range",
              "source",
              "triple_id"
            ],
            "title": "By",
            "description": "How to select what to forget."
          },
          "scope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            ],
            "title": "Scope",
            "description": "The target: entity name, [start, end] timestamps, source string, or triple ID."
          }
        },
        "type": "object",
        "required": [
          "by",
          "scope"
        ],
        "title": "ForgetRequest",
        "description": "Delete memories by entity, time range, source, or triple ID."
      },
      "ForgetResponse": {
        "properties": {
          "tombstones_emitted": {
            "type": "integer",
            "title": "Tombstones Emitted"
          }
        },
        "type": "object",
        "required": [
          "tombstones_emitted"
        ],
        "title": "ForgetResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ProactiveResponse": {
        "properties": {
          "insights": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Insights"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "default": 0
          }
        },
        "type": "object",
        "title": "ProactiveResponse"
      },
      "ProfileResponse": {
        "properties": {
          "profile": {
            "additionalProperties": true,
            "type": "object",
            "title": "Profile"
          }
        },
        "type": "object",
        "required": [
          "profile"
        ],
        "title": "ProfileResponse"
      },
      "ReconstructRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "description": "What to reconstruct, e.g. 'What's going on with Sam?'"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "ReconstructRequest",
        "description": "Reconstruct the current living state of a situation or person."
      },
      "ReconstructResponse": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Narrative reconstruction or refusal."
          },
          "refusal": {
            "type": "boolean",
            "title": "Refusal",
            "default": false
          },
          "grounding": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Grounding"
          },
          "edge_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Edge Ids"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ReconstructResponse"
      },
      "RetrieveRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "description": "The question to answer from stored memory."
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "RetrieveRequest",
        "description": "Answer a question from continuity memory."
      },
      "RetrieveResponse": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "description": "The answer, or a refusal message."
          },
          "refusal": {
            "type": "boolean",
            "title": "Refusal",
            "description": "True if the information was not found.",
            "default": false
          },
          "grounding": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Grounding",
            "description": "Source texts that ground the answer."
          },
          "edge_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Edge Ids"
          },
          "return_field": {
            "type": "string",
            "title": "Return Field",
            "default": "episodic"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "RetrieveResponse"
      },
      "StoreRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "description": "The user's message to store."
          },
          "speaker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker",
            "description": "The user's name. 'I' resolves to this."
          },
          "source_timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Timestamp",
            "description": "ISO 8601 datetime of the message."
          },
          "model_response": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Response",
            "description": "The AI's response text."
          },
          "llm_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Id",
            "description": "Which LLM is calling (e.g. 'chatgpt')."
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence",
            "description": "STT confidence, 0-1.",
            "default": 0.9
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "StoreRequest",
        "description": "Store a message into continuity memory."
      },
      "StoreResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "StoreResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:7130"
    }
  ]
}